feat: 동일한 기능을 하는 hook 통합
